+++ src/Compilers/Core/Portable/PEWriter/MetadataWriter.cs
@@ -1237,7 +1237,7 @@ namespace Microsoft.Cci
             try
             {
                 writer.WriteByte((byte)'.');
-                writer.WriteCompressedInteger((uint)permissionSet.Length);
+                writer.WriteCompressedInteger(permissionSet.Length);
                 this.SerializePermissionSet(permissionSet, writer);
                 result = metadata.GetBlob(writer);
             }
@@ -3284,15 +3284,15 @@ namespace Microsoft.Cci
 
         private void SerializeMarshallingDescriptor(IMarshallingInformation marshallingInformation, BlobBuilder writer)
         {
-            writer.WriteCompressedInteger((uint)marshallingInformation.UnmanagedType);
+            writer.WriteCompressedInteger((int)marshallingInformation.UnmanagedType);
             switch (marshallingInformation.UnmanagedType)
             {
                 case UnmanagedType.ByValArray: // NATIVE_TYPE_FIXEDARRAY
                     Debug.Assert(marshallingInformation.NumberOfElements >= 0);
-                    writer.WriteCompressedInteger((uint)marshallingInformation.NumberOfElements);
+                    writer.WriteCompressedInteger(marshallingInformation.NumberOfElements);
                     if (marshallingInformation.ElementType >= 0)
                     {
-                        writer.WriteCompressedInteger((uint)marshallingInformation.ElementType);
+                        writer.WriteCompressedInteger((int)marshallingInformation.ElementType);
                     }
 
                     break;
@@ -3329,20 +3329,20 @@ namespace Microsoft.Cci
 
                 case UnmanagedType.LPArray: // NATIVE_TYPE_ARRAY
                     Debug.Assert(marshallingInformation.ElementType >= 0);
-                    writer.WriteCompressedInteger((uint)marshallingInformation.ElementType);
+                    writer.WriteCompressedInteger((int)marshallingInformation.ElementType);
                     if (marshallingInformation.ParamIndex >= 0)
                     {
-                        writer.WriteCompressedInteger((uint)marshallingInformation.ParamIndex);
+                        writer.WriteCompressedInteger(marshallingInformation.ParamIndex);
                         if (marshallingInformation.NumberOfElements >= 0)
                         {
-                            writer.WriteCompressedInteger((uint)marshallingInformation.NumberOfElements);
+                            writer.WriteCompressedInteger(marshallingInformation.NumberOfElements);
                             writer.WriteByte(1); // The parameter number is valid
                         }
                     }
                     else if (marshallingInformation.NumberOfElements >= 0)
                     {
                         writer.WriteByte(0); // Dummy parameter value emitted so that NumberOfElements can be in a known position
-                        writer.WriteCompressedInteger((uint)marshallingInformation.NumberOfElements);
+                        writer.WriteCompressedInteger(marshallingInformation.NumberOfElements);
                         writer.WriteByte(0); // The parameter number is not valid
                     }
 
@@ -3351,7 +3351,7 @@ namespace Microsoft.Cci
                 case UnmanagedType.SafeArray:
                     if (marshallingInformation.SafeArrayElementSubtype >= 0)
                     {
-                        writer.WriteCompressedInteger((uint)marshallingInformation.SafeArrayElementSubtype);
+                        writer.WriteCompressedInteger((int)marshallingInformation.SafeArrayElementSubtype);
                         var elementType = marshallingInformation.GetSafeArrayElementUserDefinedSubtype(Context);
                         if (elementType != null)
                         {
@@ -3362,7 +3362,7 @@ namespace Microsoft.Cci
                     break;
 
                 case UnmanagedType.ByValTStr: // NATIVE_TYPE_FIXEDSYSSTRING
-                    writer.WriteCompressedInteger((uint)marshallingInformation.NumberOfElements);
+                    writer.WriteCompressedInteger(marshallingInformation.NumberOfElements);
                     break;
 
                 case UnmanagedType.Interface:
@@ -3370,7 +3370,7 @@ namespace Microsoft.Cci
                 case UnmanagedType.IUnknown:
                     if (marshallingInformation.IidParameterIndex >= 0)
                     {
-                        writer.WriteCompressedInteger((uint)marshallingInformation.IidParameterIndex);
+                        writer.WriteCompressedInteger(marshallingInformation.IidParameterIndex);
                     }
 
                     break;
@@ -3452,7 +3452,7 @@ namespace Microsoft.Cci
                 var customAttributeArgsBuilder = PooledBlobBuilder.GetInstance();
                 var namedArgsEncoder = new BlobEncoder(customAttributeArgsBuilder).PermissionSetArguments(customAttribute.NamedArgumentCount);
                 SerializeCustomAttributeNamedArguments(namedArgsEncoder, customAttribute);
-                writer.WriteCompressedInteger((uint)customAttributeArgsBuilder.Count);
+                writer.WriteCompressedInteger(customAttributeArgsBuilder.Count);
 
                 customAttributeArgsBuilder.WriteContentTo(writer);
                 customAttributeArgsBuilder.Free();
@@ -3673,7 +3673,7 @@ namespace Microsoft.Cci
             return encoder.EndModifiers();
         }
 
-        private uint GetNumberOfInheritedTypeParameters(ITypeReference type)
+        private int GetNumberOfInheritedTypeParameters(ITypeReference type)
         {
             INestedTypeReference nestedType = type.AsNestedTypeReference;
             if (nestedType == null)
@@ -3687,7 +3687,7 @@ namespace Microsoft.Cci
                 nestedType = specializedNestedType.UnspecializedVersion;
             }
 
-            uint result = 0;
+            int result = 0;
             type = nestedType.GetContainingType(Context);
             nestedType = type.AsNestedTypeReference;
             while (nestedType != null)
